Revert "Merge remote-tracking branch 'upstream/master'"
This reverts commit edb83c76c5c341871272ec93d2df8bbfbc79bc6f, reversing
changes made to ede08d1fc16045f423c5b44a09f0ec98101c3d7b.

+++ src/main/java/io/github/addoncommunity/galactifun/base/aliens/Firestorm.java
@@ -23,7 +23,7 @@ import net.kyori.adventure.text.serializer.plain.PlainTextComponentSerializer;
 public final class Firestorm extends Alien<Blaze> implements Listener {
 
 
-    public Firestorm(@NonNull String id, @NonNull String name, double maxHealth, double spawnChance) {
+    public Firestorm(@NonNull String id, @NonNull String name, double maxHealth, int spawnChance) {
         super(Blaze.class, id, name, maxHealth, spawnChance);
 
         Events.registerListener(this);
